requirements.txt

torch
transformers
sentence-transformers
faiss-cpu
numpy
tqdm
